# Use while loop for "CapitalizedWordsNamingConvention.toTokens" instead of regex to improve performance