Thanks for your work on PureConfig, it's a great lib!

We use PureConfig to read and write quite big configs (~7MB), so I checked the performance the other day. 

For testing purposes, I used an app that in a loop performs:
```scala
val rendered = implicitly[ConfigWriter[AppConfig]].to(appConfig).render(ConfigRenderOptions.concise())
implicitly[ConfigReader[AppConfig]].from(ConfigFactory.parseString(rendered).root()) match {
  case Left(error) => throw new IllegalStateException(s"Invalid config: $error")
  case Right(v) => v
}
```

It seems that the majority of time is spent in regex that is used by `CapitalizedWordsNamingConvention.toTokens`:

<img width="1751" alt="pureconfig-master" src="https://user-images.githubusercontent.com/7759586/122118080-ee2e7300-ce27-11eb-947e-1036aab65e5c.png">

and here's a fragment of a Flame Graph:

<img width="1462" alt="pureconfig-master-flamegraph" src="https://user-images.githubusercontent.com/7759586/122119114-1e2a4600-ce29-11eb-8397-36ec3e51c60a.png">

This PR rewrites `CapitalizedWordsNamingConvention.toTokens` to use while loop instead of regex. After the change, "own time" of `CapitalizedWordsNamingConvention.toTokens` is over 8x lower:

<img width="1753" alt="image" src="https://user-images.githubusercontent.com/7759586/122120022-30f14a80-ce2a-11eb-85a4-165a556350e6.png">

and similarly, here's fragment of a Flame Graph:

<img width="1453" alt="pureconfig-fix-flamegraph" src="https://user-images.githubusercontent.com/7759586/122122612-4fa51080-ce2d-11eb-88e1-c524d2871d1b.png">

When it comes to reading our real-life config from JSON and writing that config to JSON, here are the results before and after the change:
| Type | Avg Time |
|------|------|
| Writing config before | 458ms |
| Reading config before | 905ms |
| Writing config after | 177ms |
| Reading config after | 522ms |

For our use case, the change means that reading configs from JSON to a case class representation is ~40% faster and writing configs from case class to JSON is ~60% faster.