@leifwickland 2021-06-16
In principle, I like this change. Thanks for the contribution, @pwliwanow.
Before approving it, I'd like to see:
- Attempt to convert the looping to a tailrec function to eliminate the explicit `var`s.
- A new (possibly transient) unit test that generates thousands of random strings and runs them through both implementations, verifying both produce the same chunks. Perhaps it'd be sensible to commit the input and expected output to a file and have that become an ongoing test of this new algorithm.