@jcazevedo 2021-06-16
I'm a bit wary of moving this to a noticeable more complicated implementation. Performance isn't something we're super focused on in PureConfig, and I don't think these particular performance gains justify the change. The most typical use case of PureConfig is loading configurations at the start of an application, and at that point, the extra 400ms don't seem very troublesome to me.

For situations where performance at this point is relevant, we provide ways to customize the `ConfigFieldMapping` used in derived readers and writers (https://pureconfig.github.io/docs/overriding-behavior-for-case-classes.html#field-mappings) so you can have a more performant custom implementation for your application.