@ruippeixotog 2021-06-16
+1 for what @jcazevedo said. In many cases (and even more in functional programming) performance is usually traded off for code conciseness, depending on how low-level the library is supposed to sit. In PureConfig performance is definitely not a core requirement and switching a 2 LOC implementation with a 50+ LOC implementation of a FSM is quite an extreme case of losing conciseness and readability.

As @jcazevedo mentioned, this is a perfect use case for you to have your own implementation of `ConfigFieldMapping` and use it by exporting product and coproduct hints. However, I believe that between a regex and a full FSM there are a lot of intermediate solutions that produce faster results without sacrificing on readability (e.g. by folding the string?).