@pwliwanow 2021-06-18
Understandable, we'll customize `ConfigFieldMapping` then